requirements.txt
streamlit
numpy
pandas
scikit-learn
tensorflow
python-chess
joblib